pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "locus"
version = "0.1.0"
description = "Python-first backend foundation for the Locus map workbench"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.17.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.10.0",
  "redis>=6.4.0",
  "sqlalchemy>=2.0.40",
  "uvicorn>=0.35.0",
]

[project.scripts]
locus-api = "app.main:run"

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "ruff>=0.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
